feat(api): migrate method from prescription context
